requirements.txt
python-twitch-client
requests
